==> notebooks/1.api_diet.ipynb <==
import requests
import json
import pandas as pd
import sys
import os
sys.path.append(os.path.abspath('../'))
from src import soporte_dietas as sd
import dotenv 
dotenv.load_dotenv()
key = os.getenv("key")


def obtener_rutinas(api_key, num_rutinas=1000):
    """
    Obtiene un número específico de rutinas de ejercicio de una API externa.
    
    Args:
        api_key (str): La clave de API para autenticar la solicitud.
        num_rutinas (int): El número de rutinas a obtener. El valor por defecto es 1000.

    Returns:
        list: Una lista de diccionarios que representan las rutinas obtenidas de la API.
    """
    # URL de la API para obtener ejercicios
    url = "https://exercises-by-api-ninjas.p.rapidapi.com/v1/exercises"
    
    
    headers = {
        "x-rapidapi-key": api_key,  # La clave de API proporcionada
        "x-rapidapi-host": "exercises-by-api-ninjas.p.rapidapi.com"  # Host de la API
    }

    todas_rutinas = []  
    
    # Bucle para obtener el número deseado de rutinas
    for i in range(num_rutinas):
        # Parámetros de consulta para filtrar rutinas de tipo "cardio" y dificultad "intermediate"
        querystring = {"type": "cardio", "difficulty": "intermediate"}
        
        # Realiza la solicitud GET a la API
        response = requests.get(url, headers=headers, params=querystring)
        
        # Verifica si la respuesta es exitosa
        if response.status_code == 200:
            # Si la respuesta es correcta, se convierte el JSON a un objeto de Python
            data = response.json()
            # Se agrega la nueva información a la lista de todas las rutinas
            todas_rutinas.extend(data)
            
        else:
            print(f"Error en la rutina {i + 1}: {response.status_code}")
            print(response.text)
            break  

    return todas_rutinas 


df= pd.todas_rutinas                

nombres_columnas = {
    'name': 'nombre',
    'type': 'tipo',
    'muscle': 'músculo',
    'equipment': 'equipo',
    'difficulty': 'dificultad',
    'instructions': 'instrucciones'}

df.rename(columns=nombres_columnas, inplace=True)                   #renombramos las columnas

df.to_csv('../datos/rutinas.csv', index=False)              #guardamos a csv
