==> notebooks/2.api_recetas.ipynb <==
import requests
import json
import pandas as pd
import sys
import os
sys.path.append(os.path.abspath('../'))
from src import soporte_Re as sr
import dotenv 
dotenv.load_dotenv()
key = os.getenv("key")


url = "https://tasty.p.rapidapi.com/recipes/list"

querystring = {
    "from":"0","size":"100","q":"protein"}
    
headers = {
        "x-rapidapi-key": key,
        "x-rapidapi-host": "tasty.p.rapidapi.com"
    }

response = requests.get(url, headers=headers, params=querystring)
data = response.json()

results = data['results']

df= sr.recetas(results)


df.head(2)


df.head(2)
df['numero de receta'] = df['recipe_id'].str.split(':').str[-1]

df.head()
df.drop(columns=['recipe_id','approved_at','aspect_ratio','brand_id','created_at'])

nombres_columnas = { 'beauty_url': 'url_belleza', 'brand': 'marca', 'buzz_id': 'id_buzz', 'compilation_id': 'id_compilación', 'compilation_name': 'nombre_compilación', 'compilation_description': 'descripción_compilación', 'language': 'idioma', 'name': 'nombre', 'slug': 'slug', 'thumbnail_url': 'url_miniatura', 'video_url': 'url_video', 'yields': 'rinde', 'numero de receta': 'numero_de_receta' } 
df.rename(columns=nombres_columnas, inplace=True)

columnas_a_eliminar = [ 'recipe_id', 'approved_at', 'aspect_ratio', 'url_belleza', 'marca', 'brand_id', 'id_buzz', 'id_compilación' ] 
df.drop(columns=columnas_a_eliminar, inplace=True)

columnas_a_eliminar = [ 'created_at', 'created_at'] 
df.drop(columns=columnas_a_eliminar, inplace=True)

df

df.to_csv('../datos/recetas_limpias.csv', index=False)
