==> notebooks/3.ScrapProzis.ipynb <==
from bs4 import BeautifulSoup
import requests
import pandas as pd
import numpy as np
from time import sleep
from selenium import webdriver  
from selenium.webdriver.common.by import By
import re
import sys
import os
sys.path.append(os.path.abspath('../'))
from src import soporte_scrap as sc
import dotenv 
dotenv.load_dotenv()

driver = webdriver.Chrome()
url = 'https://www.pontemasfuerte.com'
driver.get(url)
driver.maximize_window()                                    #maximizo ventana
driver.find_element(By.CSS_SELECTOR, '#button1').click()        #boton cookies
sleep(5)
driver.find_element(By.XPATH, '//*[@id="vue-welcome-modal"]/div/div/header/p[2]').click()           #click ventana de cabecera
sleep(5)
driver.find_element(By.CSS_SELECTOR, '#pmfmenubar > div > div > ul > li:nth-child(1) > a').click()  #click todos los productos
#click para ver mas productos, podempos darle las veces que queramos...
driver.find_element(By.CSS_SELECTOR, '#wrapper > main > div.row.product-list > div > div.row > div.medium-8.large-10.columns > div:nth-child(6) > div > div > div').click()
sleep(5)
driver.find_element(By.CSS_SELECTOR, '#wrapper > main > div.row.product-list > div > div.row > div.medium-8.large-10.columns > div:nth-child(6) > div > div > div > span').click()
sleep(5)
driver.find_element(By.CSS_SELECTOR, '#wrapper > main > div.row.product-list > div > div.row > div.medium-8.large-10.columns > div:nth-child(6) > div > div > div > span').click()
sleep(5)
driver.find_element(By.CSS_SELECTOR, '#wrapper > main > div.row.product-list > div > div.row > div.medium-8.large-10.columns > div:nth-child(6) > div > div > div > span').click()

html_page = driver.page_source
sopa = BeautifulSoup(html_page)

Sopa_produc = sopa.findAll('div', class_ = 'item-lista not_mobile')


lista_productos = sc.scrap_fuerte(Sopa_produc)


lista_productos[1]

lista_productos
df = pd.DataFrame(lista_productos, columns=["Nombre y peso", "Precio", "cantidad de reseñas","decripcion",'rating'])

df[['Nombre', 'Peso']] = df['Nombre y peso'].str.extract(r'(.+)\s(\d+\s\w+)') 
#Eliminamos  la columna original 'Nombre y peso' 
df = df.drop(columns=['Nombre y peso'])

df.head()

df['cantidad de reseñas'] = df['cantidad de reseñas'].str.replace(r'[()]', '', regex=True)

df['Precio'] = df['Precio'].astype(float) 
df['cantidad de reseñas'] = df['cantidad de reseñas'].astype(float) 
df['rating'] = df['rating'].astype(float)

df.head(1)

df[['Peso', 'Unidad']] = df['Peso'].str.extract(r'(\d+)\s*(\w+)')
df['Peso'] = df['Peso'].astype(float)

df['Peso'] = df['Peso'].fillna(0)

df

df.to_csv('../datos/datos_scrap_fuerte.csv', index=False)
