==> notebooks/4.scrap_super.ipynb <==
from bs4 import BeautifulSoup
import requests
import pandas as pd
import numpy as np
from time import sleep
from selenium import webdriver  
from selenium.webdriver.common.by import By
from webdriver_manager.chrome import ChromeDriverManager 
from selenium.webdriver.common.keys import Keys  
from selenium.webdriver.support.ui import Select  
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.common.exceptions import NoSuchElementException 
import dotenv 
dotenv.load_dotenv()
import sys
import os
sys.path.append(os.path.abspath('../'))
import re
from src import soporte_scrap as sc

driver = webdriver.Chrome()
url = 'https://www.ahorramas.com/'
driver.get(url)
driver.maximize_window()         
sleep(2)
driver.find_element(By.XPATH, '/html/body/div[7]/div[2]/div/div/div[2]/div/div/button[2]').click()        #boton cookies 
sleep(5)
driver.find_element(By.CSS_SELECTOR, '#frescos > span').click()   #productos frescos
sleep(5)
driver.find_element(By.CSS_SELECTOR, '#product-search-results > div:nth-child(2) > div.col-sm-12.col-lg-9 > div.row.product-grid > div.col-12.grid-footer > div > div > button').click()          #mas resultados
sleep(5)
driver.find_element(By.CSS_SELECTOR, '#product-search-results > div:nth-child(2) > div.col-sm-12.col-lg-9 > div.row.product-grid > div.col-12.grid-footer > div > div > button').click()          #mas resultados
html_page = driver.page_source
sopa = BeautifulSoup(html_page)

sopa_productos = sopa.findAll('div', class_ = 'tile-body')


lista_productos=sc.productos_frescos(sopa_productos)

lista_productos
df = pd.DataFrame(lista_productos, columns=["Nombre y peso", "Precio", "Descuentos",'rango de precios bajada'])


df.head(2)

df['Precio'] = df['Precio'].str.replace('\n', '').str.replace('€', '').str.replace(',', '.').astype(float)  #precio a float
df['rango de precios bajada'] = df['rango de precios bajada'].str.replace('\n', '')


df['rango de precios bajada'] = df['rango de precios bajada'].str.replace('\n', '').str.replace('/','').str.replace('kg','')

df[['precio_actual', 'precio_sin_descuento']] = df['rango de precios bajada'].str.extract(r'\(([\d,]+)€Kg\)\s*\|\s*\(([\d,]+)€Kg\)')
df['precio_sin_descuento'] = df['precio_sin_descuento'].fillna(0)
df['precio_actual'] = df['precio_actual'].fillna(0)
df = df.drop(columns=['rango de precios bajada'])

df['precio_sin_descuento'] = df['precio_sin_descuento'].str.replace(',','.').astype(float)
df['precio_actual'] = df['precio_actual'].str.replace(',','.').astype(float)

df['Descuentos'] = df['Descuentos'].str.replace('%','').str.replace(',','.').astype(float)

df

df.to_csv('../datos/datos_scrap_ahorra.csv', index=False)
