==> notebooks/5.Tablas_e_insercion.ipynb <==
import requests
import pandas as pd
import numpy as np
import sys
import os
import psycopg2
from psycopg2 import OperationalError, errorcodes, errors 

sys.path.append(os.path.abspath('../'))
from src import soporte_tablas as st

conexion = st.conexion("Salud")
cursor = conexion.cursor()

st.crear_tablas(conexion)

df_productos_precios = pd.read_csv('../datos/datos_scrap_ahorra.csv')
df_productos_fit = pd.read_csv('../datos/datos_scrap_fuerte.csv')
df_recetas = pd.read_csv('../datos/recetas_limpias.csv')
df_ejercicios = pd.read_csv('../datos/rutinas.csv')


df_productos_fit.head()         #reordeno el df puesto que hemos creado la tabla y la insercion asi
df_productos_fit = df_productos_fit[['Nombre', 'Precio', 'cantidad de reseñas', 'decripcion', 'rating', 'Peso', 'Unidad']]


df_productos_precios.head().dtypes
df_productos_precios['Descuentos'] = df_productos_precios['Descuentos'].astype(float)
df_productos_precios['precio_actual'] = df_productos_precios['precio_actual'].astype(float) 
df_productos_precios['precio_sin_descuento'] = df_productos_precios['precio_sin_descuento'].astype(float)
df_productos_precios['Precio'] = df_productos_precios['Precio'].astype(float)

df_recetas = df_recetas.rename(columns={ 'nombre_compilación': 'nombre_compilacion', 'descripción_compilación':'descripcion_compilacion' })

df_ejercicios.head()
df_ejercicios = df_ejercicios.rename(columns={ 'músculo': 'musculo'})


df_productos_fit
df_productos_fit['tipo_producto'] = 'suplemento'

st.insertar_datos(df_productos_fit, df_productos_precios, df_recetas, df_ejercicios, conexion)
